# swagger.yaml
swagger: '2.0'
info:
  description: "App - TecWeb - Sistema desenvolvido ao longo da disciplina de TecWeb-UfSJ"
  version: "1.0.0"
  title: "Patinete Elétricos"
  contact:
    email: "ygormagalhaesrc@gmail.com"
host: "Patinete.ccomp.ufsj.edu.br"
basePath: "/v1"
tags:
- name: "patinete"
  description: "Informação Sobre Os patinetes Eletricos"
- name: "pontoDosPatinetes"
  description: "Informaçao da localidade de onde alugar os patinetes"
- name: "usuarios"
  description: "Operações gerenciamento dos usuarios"
- name: "admin"
  description: "Painel de gerenciar os patinetes, localidade e usuarios"
paths:
  /patinete:
    post:
      tags:
      - "patinete"
      summary: "Adiciona um novo Patinete eletrico no sistema de aluguel"
      description: "addPatinete"
      consumes:
      - "applications/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "objtedo bicicleta a ser adicionado no sistema"
        required: true
        schema:
          $ref: "#/definitions/Patinete"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Patinete"
  /patinete/{patineteId}:
    get:
      tags:
      - "patinete"
      summary: "Busca Patinete pelo Id"
      description: "getPatinete"
      produces:
      - "application/json"
      parameters:
      - name: "patineteId"
        in: "path"
        description: "ID do patinete retornado"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "sucesso na operação"
          schema:
            $ref: "#/definitions/Patinete"
        400:
          description: "invalido Id passado"
        404:
          description: "Patinete não encontrado"
    put:
      tags:
      - "patinete"
      summary: "Atualizar patinete no pontoDosPatinetes"
      description: "putPatinete"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "patineteId"
        in: "path"
        description:   "ID do patinete que será atualizado  "
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "atualizar objeto patinete"
        schema:
          $ref: "#/definitions/Patinete"
      responses:
        200:
          description: "sucesso na operação"
          schema:
            $ref: "#/definitions/Patinete"
        400:
          description: "invalido Id passado"
        404:
          description: "Patinete não encontrado"
    delete:
      tags:
      - "patinete"
      summary: "deletar patinete no pontoDosPatinetes"
      operationId: "deletePatinete"
      produces:
      - "application/json"
      parameters:
      - name: "patineteId"
        in: "path"
        description:   "ID do patinete que será deletada  "
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "sucesso na operação"
          schema:
            $ref: "#/definitions/Patinete"
        400:
          description: "invalido Id passado"
        404:
          description: "Patinete não encontrado" 
      
definitions:
  Patinete:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      marca:
        type: "string"
      modelo:
        type: "string"
      status:
        type: "string"
        enum:
        - "disponivel"
        - "alugada"
      